% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGDINA.stEM.R
\name{autoGDINA.stEM}
\alias{autoGDINA.stEM}
\title{Stochastic EM Algorithm for High-Dimensional Cognitive Diagnosis Models}
\usage{
autoGDINA.stEM(
  dat,
  Q,
  item.parm = NULL,
  maxitr = 100,
  eps1 = 2,
  eps2 = 0.4,
  frac1 = 0.1,
  frac2 = 0.5
)
}
\arguments{
\item{dat}{A required \eqn{N \times J} data matrix of N examinees to J items. Missing values are currently not allowed.}

\item{Q}{A required binary item and attribute association matrix.A required \eqn{J \times K} item-attribute association matrix, where J is the number of items and K is the number of attributes.}

\item{item.parm}{A initial item parameter estimates.}

\item{maxitr}{The maximum iterations allowed.}

\item{eps1}{A positive convergence criterion used during the burn-in stage of the stochastic EM algorithm.}

\item{eps2}{A positive convergence criterion used after the burn-in.}

\item{frac1}{Proportion of the first part of the Markov chain used in the Geweke diagnostic.}

\item{frac2}{Proportion of the last part of the Markov chain used in the Geweke diagnostic.}
}
\value{
a list with elements
\describe{
\item{catprob.parm}{estimated item category response probabilities}
\item{ip}{estimated final parameter obtained by averaging from each batches}
\item{alpha}{estimated attribute mastery profiles}
\item{alpha.list}{latent attribute profile estimates from each batches}
\item{total.number.of.batch}{total number of retained batches used for estimation}
\item{final.chain}{final length of the Markov chain}
\item{burn.in.size}{the number of burn-in iterations discarded}
\item{plist}{estimated item parameters from each batch}
}
}
\description{
This function estimates the model parameters of CDM using the stochastic EM algorithm with sequential Gibbs sampling.
}
\examples{
\dontrun{
dat <- realdata_ECPE$dat
Q <- realdata_ECPE$Q
fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
fit
CA(fit)
}
}
\author{
Wenchao Ma, The University of Minnesota, \email{wma@umn.edu}
}
