% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.var.R
\name{batch.var}
\alias{batch.var}
\title{Batch variance calculation}
\usage{
batch.var(plist, n)
}
\arguments{
\item{plist}{A list of parameter estimates collected across batches.}

\item{n}{Number of batches}
}
\value{
A vector of batch variance estimates.
}
\description{
Computes variance of parameter estimates across batches in the stochastic EM algorithm.
}
\author{
Wenchao Ma, The University of Minnesota, \email{wma@umn.edu}
}
