% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPIP.R
\docType{data}
\name{IPIP}
\alias{IPIP}
\title{International Personality Item Pool (IPIP) data}
\format{
A list of responses and Q-matrix with components:
\describe{
\item{\code{dat}}{Responses of 4271 participants to 120 items.}
\item{\code{Q}}{The \eqn{120 \times 30} Q-matrix.}
}
}
\usage{
IPIP
}
\description{
The IPIP-NEO-120 data are based on the public-domain personality inventory
developed from the International Personality Item Pool (IPIP; Johnson, 2014).
}
\details{
The data consist of responses from 4271 participants to 120 dichotomized items
and the corresponding Q-matrix relating the 120 items to 30 personality facets.
}
\examples{
\dontrun{
data(IPIP)
str(IPIP)
fit <- autoGDINA.stEM(
  dat = IPIP$dat,
  Q = IPIP$Q
  )
}

}
\references{
Johnson, J. A. (2014). Measuring thirty facets of the five factor model with a 120-item  public domain inventory: Development of the ipip-neo-120. Journal of Research in  Personality, 51, 78–89.
}
\author{
Yuxuan Mei, The University of Minnesota, \email{mei00060@umn.edu}
}
\keyword{datasets}
