% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGDINA.stEM.R
\name{stEM}
\alias{stEM}
\title{Stochastic EM Algorithm for High-Dimensional Cognitive Diagnosis Models}
\usage{
stEM(
  dat,
  Q,
  item.parm = NULL,
  maxitr = 100,
  eps1 = 2,
  eps2 = 0.4,
  frac1 = 0.1,
  frac2 = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{dat}{A required \eqn{N \times J} data matrix of N examinees to J items. Missing values are not currently supported.}

\item{Q}{A required \eqn{J \times K} item-attribute association matrix, where J is the number of items and K is the number of attributes.}

\item{item.parm}{Optional initial item parameter estimates.}

\item{maxitr}{Maximum number of iterations.}

\item{eps1}{A positive convergence criterion used during the burn-in stage of the stochastic EM algorithm.}

\item{eps2}{A positive convergence criterion used after the burn-in.}

\item{frac1}{Proportion of the first part of the Markov chain used in the Geweke diagnostic.}

\item{frac2}{Proportion of the last part of the Markov chain used in the Geweke diagnostic.}

\item{verbose}{Logical; if TRUE, prints progress messages during the estimation process.}
}
\value{
a list with elements
\describe{
\item{catprob.parm}{estimated item category response probabilities}
\item{ip}{estimated final item parameters obtained by averaging across retained batches}
\item{alpha}{estimated attribute mastery profiles}
\item{alpha.list}{attribute mastery profile estimates from each retained batch}
\item{total.number.of.batch}{total number of retained batches used for estimation}
\item{final.chain}{final length of the Markov chain}
\item{burn.in.size}{the number of burn-in iterations discarded}
\item{plist}{estimated item parameters from each retained batch}
}
}
\description{
Estimates model parameters for cognitive diagnosis models using the stochastic expectation-maximization (stEM) algorithm with sequential Gibbs sampling.
}
\examples{
\donttest{
dat <- simN2000K20$data
Q <- simN2000K20$Q
fit <- stEM(dat = dat, Q = Q)
fit
}
}
