% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.var.R
\name{batch.var}
\alias{batch.var}
\title{Batch variance calculation}
\usage{
batch.var(plist, n)
}
\arguments{
\item{plist}{A list of parameter estimates collected across batches.}

\item{n}{Number of batches}
}
\value{
A vector of batch variance estimates.
}
\description{
Internal function used by autoGDINA.stEM.
}
\examples{
\dontrun{
dat <- realdata_ECPE$dat
Q <- realdata_ECPE$Q
fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
fit
CA(fit)



}
}
\author{
Wenchao Ma, The University of Minnesota, \email{wma@umn.edu}
}
