% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel}
\alias{kernel}
\title{Gibbs Sampling Kernel for stEM Algorithm}
\usage{
kernel(dat, Q, J, reduced.profiles, alpha, B, ip.only = TRUE)
}
\arguments{
\item{dat}{A required \eqn{N \times J} data matrix of N examinees to J items. Missing values are currently not allowed.}

\item{Q}{A required \eqn{J \times K} item-attribute association matrix, where J is the number of items and K is the number of attributes.}

\item{J}{The number of items.}

\item{reduced.profiles}{A list of reduced attribute profiles for each item.}

\item{alpha}{The initial latent attribute profiles.}

\item{B}{The number of Gibbs sampling iterations.}

\item{ip.only}{If \code{TRUE}, only item parameter estimates are returned.}
}
\value{
a list with elements
\describe{
\item{ip}{estimated item parameters from Gibbs sampling iterations}
\item{alpha}{latent attribute profiles from the final Gibbs sampling iteration}
\item{a}{saved latent attribute profiles across Gibbs sampling iterations}
\item{item.parm}{final item parameter estimates}
}
}
\description{
Performs Gibbs sampling updates of attribute profiles and item parameters within the stochastic EM algorithm.
}
\author{
Wenchao Ma, The University of Minnesota, \email{wma@umn.edu}
}
